refactor: mover la carpeta componentes fuera de playlist
